Moved audio permissions request to the new entry view on clicking the new audio attachment button
