+++ app/src/main/java/com/cabbage/sdpjournal/EntryViewActivity.java
@@ -44,7 +44,6 @@ public class EntryViewActivity extends AppCompatActivity implements View.OnClick
 
         firebaseAuth = FirebaseAuth.getInstance();
         init();
-        requestPermissions();
     }
 
     private void init() {
@@ -216,46 +215,6 @@ public class EntryViewActivity extends AppCompatActivity implements View.OnClick
         i.putExtra("filter", getIntent().getStringExtra("filter"));
         i.putExtra("search", getIntent().getStringExtra("search"));
         setResult(RESULT_OK, i);
-    }
-
-    public void requestPermissions() {
-        boolean requestAudio = false;
-        boolean requestWriteExternalStorage = false;
-        if (ContextCompat.checkSelfPermission(EntryViewActivity.this, android.Manifest.permission.RECORD_AUDIO) != PackageManager.PERMISSION_GRANTED) {
-            requestAudio = true;
-        }
-
-        if (ContextCompat.checkSelfPermission(EntryViewActivity.this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
-            requestWriteExternalStorage = true;
-        }
-
-        if (requestAudio && requestWriteExternalStorage) {
-            ActivityCompat.requestPermissions(EntryViewActivity.this, new String[]{android.Manifest.permission.RECORD_AUDIO, android.Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
-        } else if (requestAudio) {
-            ActivityCompat.requestPermissions(EntryViewActivity.this, new String[]{android.Manifest.permission.RECORD_AUDIO}, 1);
-        } else if (requestWriteExternalStorage) {
-            ActivityCompat.requestPermissions(EntryViewActivity.this, new String[]{android.Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
-        switch (requestCode) {
-            case 1: {
-                if (grantResults.length == 0) {
-                    // disable the attachment function
-                } else {
-                    for (int i = 0; i < grantResults.length; i++) {
-                        if (grantResults[i] == PackageManager.PERMISSION_GRANTED) {
-                            // permission was granted, yay!
-                        } else {
-                            // disable the attachment function
-                        }
-                    }
-                    return;
-                }
-            }
-        }
     }
 
     @Override
