+++ app/src/main/java/com/cabbage/sdpjournal/NewEntryActivity.java
@@ -10,6 +10,8 @@ import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.support.annotation.NonNull;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
@@ -161,9 +163,7 @@ public class NewEntryActivity extends AppCompatActivity implements View.OnClickL
                 chooseImage();
                 return true;
             case R.id.action_audio:
-                //ask user to turn on permissions of audio and storage.
-                //if truned on ...
-                recordAudioDialog();
+                requestPermissions();
                 return true;
         }
 
@@ -406,4 +406,56 @@ public class NewEntryActivity extends AppCompatActivity implements View.OnClickL
 
         return isValid;
     }
+
+    public void requestPermissions() {
+        boolean requestAudio = false;
+        boolean requestWriteExternalStorage = false;
+        if (ContextCompat.checkSelfPermission(NewEntryActivity.this, android.Manifest.permission.RECORD_AUDIO) != PackageManager.PERMISSION_GRANTED) {
+            requestAudio = true;
+        }
+
+        if (ContextCompat.checkSelfPermission(NewEntryActivity.this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
+            requestWriteExternalStorage = true;
+        }
+
+        if (!requestAudio && !requestWriteExternalStorage) {
+            recordAudioDialog();
+        } else if (requestAudio && requestWriteExternalStorage) {
+            ActivityCompat.requestPermissions(NewEntryActivity.this, new String[]{android.Manifest.permission.RECORD_AUDIO, android.Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
+        } else if (requestAudio) {
+            ActivityCompat.requestPermissions(NewEntryActivity.this, new String[]{android.Manifest.permission.RECORD_AUDIO}, 1);
+        } else if (requestWriteExternalStorage) {
+            ActivityCompat.requestPermissions(NewEntryActivity.this, new String[]{android.Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
+        switch (requestCode) {
+            case 1: {
+                boolean allowAudio = true;
+                if (grantResults.length == 0) {
+                    // disable the attachment function
+                    allowAudio = false;
+                } else {
+                    for (int i = 0; i < grantResults.length; i++) {
+                        if (grantResults[i] == PackageManager.PERMISSION_GRANTED) {
+                            // permission was granted, yay!
+                        } else {
+                            // disable the attachment function
+                            allowAudio = false;
+                            break;
+                        }
+                    }
+                    return;
+                }
+                if (allowAudio) {
+                    recordAudioDialog();
+                }
+            }
+        }
+
+
+    }
+
 }
